+++ src/expression_cfunc.cpp
@@ -1722,9 +1722,6 @@ auto add_cfunc_impl(llvm_state &s, const std::string &name, const F &fn, std::ui
     // Restore the original insertion block.
     builder.SetInsertPoint(orig_bb);
 
-    // Run the optimisation pass.
-    s.optimise();
-
     return dc;
 }