+++ doc/changelog.rst
@@ -1,6 +1,38 @@
 Changelog
 =========
 
+1.1.0 (unreleased)
+------------------
+
+New
+~~~
+
+- It is now possible to get the LLVM bitcode of
+  an ``llvm_state``
+  (`#339 <https://github.com/bluescarni/heyoka/pull/339>`__).
+
+Changes
+~~~~~~~
+
+- The LLVM bitcode is now used internally (instead of the textual
+  representation of the IR) when copying and serialising
+  an ``llvm_state``
+  (`#339 <https://github.com/bluescarni/heyoka/pull/339>`__).
+- The optimisation pass in an ``llvm_state`` is now automatically
+  called during compilation
+  (`#339 <https://github.com/bluescarni/heyoka/pull/339>`__).
+
+Fix
+~~~
+
+- Fix the object file of an ``llvm_state`` not being
+  preserved during copy and deserialisation
+  (`#339 <https://github.com/bluescarni/heyoka/pull/339>`__).
+- Fix LLVM module name not being preserved during
+  copy and deserialisation of ``llvm_state``
+  (`#339 <https://github.com/bluescarni/heyoka/pull/339>`__).
+- Fix broken link in the docs.
+
 1.0.0 (2023-08-10)
 ------------------